toRedian ref fixed

+++ Spraky-core/x64/src/maths/maths_func.h
@@ -6,7 +6,7 @@
 namespace sparky {
 	namespace maths {
 		inline float toRadians(float degrees) {
-			return static_cast<float>(degrees * M_PI / 180.0f);
+			return (degrees * M_PI / 180.0f);
 		}
 	}
 }
